Allow Empty EventMonitors And Try To Fix `CompositeEventMonitor` Thread-Safety (#3894)
### Issue Link :link:
#3635, #3667

### Goals :soccer:
This PR attempts to fix unreproducible `CompositeEventMonitor` crashes
in two ways:

1. Passing `[]` to the `Session` inits now creates a
`CompositeEventMonitor` that's properly empty rather than always
including an `AlamofireNotifications` monitor. This should allow anyone
seeing this crash to empty out the monitors to see if the crash is
resolved. **Technically this is a breaking change** as anyone passing
`[]` currently has notification support. This should be rare, and this
fix is important enough to justify breaking that non-obvious usage.
2. The stored monitors in the `CompositeEventMonitor` are now protected
and accessed through a lock, so should be made safer. Of course, I could
never figure out how any access there could be unsafe since it should
always be accessed on the serial internal queue, so we'll see if that
helps anything.

### Implementation Details :construction:
`Session` init updated, lock added.

### Testing Details :mag:
No tests added, test suite passes.